Move util functions to util file
