+++ Apache-Arrow-Flight-Tester/common.py
@@ -5,6 +5,7 @@ from typing import Literal
 from protobuf import protocol_pb2
 from pyarrow import flight, Schema
 from pyarrow._flight import FlightInfo, ActionType, Result, Ticket
+from util import get_time_series_table_schema
 
 
 class ModelarDBFlightClient:
@@ -184,16 +185,3 @@ class ModelarDBFlightClient:
         """Return the type of the node."""
         node_type = self.do_action("NodeType", b"")
         return node_type[0].body.to_pybytes().decode("utf-8")
-
-
-def get_time_series_table_schema() -> pyarrow.Schema:
-    """Return a schema for a time series table with one timestamp column, three tag columns, and three field columns."""
-    return pyarrow.schema([
-        ("location", pyarrow.utf8()),
-        ("install_year", pyarrow.utf8()),
-        ("model", pyarrow.utf8()),
-        ("timestamp", pyarrow.timestamp("us")),
-        ("power_output", pyarrow.float32()),
-        ("wind_speed", pyarrow.float32()),
-        ("temperature", pyarrow.float32()),
-    ])
